Move toggle block comment and tests to correct layer.
