+++ src/EditorFeatures/Test/CommentSelection/ToggleBlockCommentCommandHandlerTests.cs
@@ -1,30 +1,17 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using System.Composition;
-using System.Linq;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CommentSelection;
-using Microsoft.CodeAnalysis.Editor.CSharp.CommentSelection;
 using Microsoft.CodeAnalysis.Editor.Implementation.CommentSelection;
-using Microsoft.CodeAnalysis.Editor.UnitTests;
-using Microsoft.CodeAnalysis.Editor.UnitTests.Utilities;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
-using Microsoft.CodeAnalysis.Experiments;
-using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Operations;
 using Roslyn.Test.Utilities;
-using Roslyn.Utilities;
 using Xunit;
 
-namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.CommentSelection
+namespace Microsoft.CodeAnalysis.Editor.UnitTests.CommentSelection
 {
     [UseExportProvider]
-    public class ToggleBlockCommentCommandHandlerTests
+    public class ToggleBlockCommentCommandHandlerTests : AbstractToggleBlockCommentTestBase
     {
         [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
         public void AddComment_EmptyCaret()
@@ -186,13 +173,13 @@ class C
 {
     void M()
     {
-        var/**/ i = 1;
+        va/**/r i = 1;
     }
 }";
 
             var expectedSelectedSpans = new[]
             {
-                Span.FromBounds(46, 50)
+                Span.FromBounds(45, 49)
             };
             ToggleBlockComment(markup, expected, expectedSelectedSpans);
         }
@@ -214,13 +201,13 @@ class C
 {
     void M()
     {
-        Func<int, bool> myFunc = x =>/**/ x == 5;
+        Func<int, bool> myFunc = x =/**/> x == 5;
     }
 }";
 
             var expectedSelectedSpans = new[]
             {
-                Span.FromBounds(72, 76)
+                Span.FromBounds(71, 75)
             };
             ToggleBlockComment(markup, expected, expectedSelectedSpans);
         }
@@ -322,39 +309,6 @@ class C
             var expectedSelectedSpans = new[]
             {
                 Span.FromBounds(43, 79)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void AddComment_MarkerInsideSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        [|var i = 1;
-        string s = '/*';
-        var k = 3;|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        /*var i = 1;
-        string s = '/*';
-        var k = 3;*/
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(43, 103)
             };
             ToggleBlockComment(markup, expected, expectedSelectedSpans);
         }
@@ -573,138 +527,6 @@ class C
             var expectedSelectedSpans = new[]
             {
                 Span.FromBounds(43, 105)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void AddComment_StartCommentMarkerStringInSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        [|var i = 1;
-        string s = '/*';
-        var k = 3;|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        /*var i = 1;
-        string s = '/*';
-        var k = 3;*/
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(43, 103)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void AddComment_CloseCommentMarkerStringInSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        [|var i = 1;
-        string s = '*/';
-        var k = 3;|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        /*var i = 1;
-        string s = '*/';
-        var k = 3;*/
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(43, 103)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void AddComment_CommentMarkerStringBeforeSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        string s = '/*';
-        [|var j = 2;
-        var k = 3;|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        string s = '/*';
-        /*var j = 2;
-        var k = 3;*/
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(69, 103)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void AddComment_CommentMarkerStringAfterSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        [|var i = 1;
-        var j = 2;|]
-        string s = '*/';
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        /*var i = 1;
-        var j = 2;*/
-        string s = '*/';
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(43, 77)
             };
             ToggleBlockComment(markup, expected, expectedSelectedSpans);
         }
@@ -1168,43 +990,6 @@ class C
         }
 
         [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void AddComment_DirectiveWithCommentInsideSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        [|var i = 1;
-#if false
-        /*var j = 2;*/
-#endif
-        var k = 3;|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        /*var i = 1;
-#if false
-        /*var j = 2;*/
-#endif
-        var k = 3;*/
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(43, 120),
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
         public void RemoveComment_CaretInsideBlock()
         {
             var markup =
@@ -1533,70 +1318,6 @@ class C
         }
 
         [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void RemoveComment_CommentMarkerStringNearSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        string s = '/*';
-        [|/*var i = 1;
-        var j = 2;
-        var k = 3;*/|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        string s = '/*';
-        var i = 1;
-        var j = 2;
-        var k = 3;
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(69, 119)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
-        public void RemoveComment_CommentMarkerStringInSelection()
-        {
-            var markup =
-@"
-class C
-{
-    void M()
-    {
-        [|/*string s = '/*';*/|]
-    }
-}";
-            var expected =
-@"
-class C
-{
-    void M()
-    {
-        string s = '/*';
-    }
-}";
-
-            var expectedSelectedSpans = new[]
-            {
-                Span.FromBounds(43, 59)
-            };
-            ToggleBlockComment(markup, expected, expectedSelectedSpans);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
         public void RemoveComment_SequentialBlockInSelection()
         {
             var markup =
@@ -1907,71 +1628,11 @@ class C
             ToggleBlockCommentMultiple(markup, expectedText, expectedSelectedSpans);
         }
 
-        private static void ToggleBlockComment(string markup, string expected, IEnumerable<Span> expectedSelectedSpans)
+        internal override ToggleBlockCommentCommandHandler GetToggleBlockCommentCommandHandler(TestWorkspace workspace)
         {
-            ToggleBlockCommentMultiple(markup, new string[] { expected }, new IEnumerable<Span>[] { expectedSelectedSpans });
-        }
-
-        private static void ToggleBlockCommentMultiple(string markup, string[] expectedText, IEnumerable<Span>[] expectedSelections)
-        {
-            Assert.Equal(expectedText.Length, expectedSelections.Length);
-            var exportProvider = ExportProviderCache
-                .GetOrCreateExportProviderFactory(TestExportProvider.EntireAssemblyCatalogWithCSharpAndVisualBasic.WithPart(typeof(MockToggleBlockCommentExperimentationService)))
-                .CreateExportProvider();
-            using (var workspace = TestWorkspace.CreateCSharp(markup, exportProvider: exportProvider))
-            {
-
-                var doc = workspace.Documents.First();
-                SetupSelection(doc.GetTextView(), doc.SelectedSpans.Select(s => Span.FromBounds(s.Start, s.End)));
-
-                var commandHandler = new ToggleBlockCommentCommandHandler(
+            return new ToggleBlockCommentCommandHandler(
                     workspace.ExportProvider.GetExportedValue<ITextUndoHistoryRegistry>(),
                     workspace.ExportProvider.GetExportedValue<IEditorOperationsFactoryService>());
-                var textView = doc.GetTextView();
-                var textBuffer = doc.GetTextBuffer();
-
-                for (var i = 0; i < expectedText.Length; i++)
-                {
-                    commandHandler.ExecuteCommand(textView, textBuffer, ValueTuple.Create(), TestCommandExecutionContext.Create());
-                    AssertCommentResult(doc.TextBuffer, textView, expectedText[i], expectedSelections[i]);
-                }
-            }
-        }
-
-        private static void AssertCommentResult(ITextBuffer textBuffer, IWpfTextView textView, string expectedText, IEnumerable<Span> expectedSelection)
-        {
-            Assert.Equal(expectedText, textBuffer.CurrentSnapshot.GetText());
-
-            if (expectedSelection != null)
-            {
-                AssertEx.Equal(expectedSelection, textView.Selection.SelectedSpans.Select(snapshotSpan => snapshotSpan.Span));
-            }
-        }
-
-        private static void SetupSelection(IWpfTextView textView, IEnumerable<Span> spans)
-        {
-            var snapshot = textView.TextSnapshot;
-            if (spans.Count() == 1)
-            {
-                textView.Selection.Select(new SnapshotSpan(snapshot, spans.Single()), isReversed: false);
-                textView.Caret.MoveTo(new SnapshotPoint(snapshot, spans.Single().End));
-            }
-            else if (spans.Count() > 1)
-            {
-                textView.Selection.Mode = TextSelectionMode.Box;
-                textView.Selection.Select(new VirtualSnapshotPoint(snapshot, spans.First().Start),
-                                          new VirtualSnapshotPoint(snapshot, spans.Last().End));
-                textView.Caret.MoveTo(new SnapshotPoint(snapshot, spans.Last().End));
-            }
-        }
-
-        [ExportWorkspaceService(typeof(IExperimentationService), WorkspaceKind.Test), Shared]
-        private class MockToggleBlockCommentExperimentationService : IExperimentationService
-        {
-            public bool IsExperimentEnabled(string experimentName)
-            {
-                return WellKnownExperimentNames.RoslynToggleBlockComment.Equals(experimentName);
-            }
         }
     }
 }
